[Support] Move fill operations from Kernels/Misc to Support/ML (#30666)
These operations are so primitive, self contained, and have really not
too much to do with the "traditional" kernels we write in mojo. Moving
them to support also simplifies the layering of GenericML

Part of #30587

MODULAR_ORIG_COMMIT_REV_ID: 9ebf2feff7931df66f553d733eeb91d88c70ab6d

+++ Support/unittests/DeviceSpecsTest.cpp
@@ -108,11 +108,11 @@ TEST(DeviceSpecs, ReconcileDeviceSpecs) {
     ASSERT_TRUE(itr != mapOr->end());
     // provided
     ASSERT_EQ(itr->second.first.ref.label, "cuda");
-    ASSERT_EQ(itr->second.first.ref.id, 0);
+    ASSERT_EQ(itr->second.first.ref.id, (size_t)0);
     ASSERT_EQ(itr->second.first.target.arch, "sm_80");
     // required
     ASSERT_EQ(itr->second.second.ref.label, "cudax");
-    ASSERT_EQ(itr->second.second.ref.id, 7);
+    ASSERT_EQ(itr->second.second.ref.id, (size_t)7);
     ASSERT_EQ(itr->second.second.target.arch, "sm_90");
   }
 }
